# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful; session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "202":
          description: New device detected — email verification required (Phase 2). No session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingVerificationResponse"
        "401":
          description: Invalid credentials
...
  /billing/invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, cancelled]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [billing]
      summary: Get invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [billing]
      summary: Update invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/invoices/{invoiceId}/pay:
    post:
      operationId: payInvoice
      tags: [billing]
      summary: Mark invoice as paid and generate receipt
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInvoiceBody"
      responses:
        "200":
          description: Invoice paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/daily-summary:
    get:
      operationId: getDailyBillingSummary
      tags: [billing]
      summary: Get daily billing summary and cash receipts
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily billing summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyBillingSummary"
  /billing/reconciliation:
    get:
      operationId: getBillingReconciliation
      tags: [billing]
      summary: End-of-day reconciliation (Z-report) — super_admin only
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily reconciliation report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingReconciliation"

  # SERVICES CATALOG (price book)
...
  /auth/verify-device:
    post:
      operationId: verifyDevice
      tags: [auth]
      summary: Consume a device verification token and issue a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceTokenBody"
      responses:
        "200":
          description: Device verified; session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyDeviceResponse"
        "400":
          description: Invalid, expired, or already-consumed token
        "503":
          description: Phase 2 not enabled
  /auth/wasnt-me:
    post:
      operationId: reportWasntMe
      tags: [auth]
      summary: Kill-switch — revoke all sessions and flag the device as rejected
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceTokenBody"
      responses:
        "200":
          description: Sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [revoked]
                required: [status]
        "400":
          description: Token expired or already consumed
        "503":
          description: Phase 2 not enabled
...
  /account/devices/{deviceId}:
    delete:
      operationId: revokeDevice
      tags: [auth]
      summary: Revoke a trusted device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Device revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [revoked]
                required: [status]
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        deviceUnverified:
          type: boolean
          description: "True when Phase 2 is active and the device is newly trusted but the email link has not yet been clicked (allow_unverified path for non-privileged roles)."
      required: [user]
    PendingVerificationResponse:
      type: object
      description: "Returned HTTP 202 when Phase 2 is active and a privileged-role user logs in from an unrecognised device. No session is issued. The user should check their email."
      properties:
        status:
          type: string
          enum: [pending_verification]
        message:
          type: string
      required: [status, message]
    DeviceTokenBody:
      type: object
      properties:
        token:
          type: string
          minLength: 16
          maxLength: 256
      required: [token]
    VerifyDeviceResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, doctor, nurse, front_desk, xray_staff, lab_staff]
        isActive:
          type: boolean
        isOnShift:
          type: boolean
        phone:
          type: string
          nullable: true
        addressLine:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, username, fullName, role, isActive, isOnShift, createdAt]
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        mrn:
          type: string
        idCardNumber:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        phone:
          type: string
        address:
          type: string
          nullable: true
        bloodType:
          type: string
          nullable: true
        allergies:
          type: string
          nullable: true
        emergencyContact:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, mrn, idCardNumber, fullName, dateOfBirth, gender, phone, isActive, createdAt]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        patientId:
          type: integer
        createdById:
          type: integer
        patient:
          $ref: "#/components/schemas/Patient"
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        paidAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, patientId, createdById, items, subtotal, discount, total, status, createdAt]
    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
      required: [description, quantity, unitPrice, total]

    # Request-side line item. Unlike InvoiceItem (a response shape), `total` is
    # NOT supplied by the client — the server computes it as quantity * unitPrice.
    InvoiceItemInput:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [description, quantity, unitPrice]
    CreateInvoiceBody:
      type: object
      properties:
        patientId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        discount:
          type: number
        notes:
          type: string
        markPaid:
          type: boolean
          description: >-
            Point-of-sale flow — when true the invoice is created already paid
            (status=paid, paidAt=now), skipping the separate pending → pay step.
      # NOTE: `createdById` is set server-side from the authenticated user
      # (req.user.userId) and is intentionally NOT part of the request body.
      required: [patientId, items]
    UpdateInvoiceBody:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        discount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        notes:
          type: string
    PayInvoiceBody:
      type: object
      properties:
        amountReceived:
          type: number
      required: [amountReceived]
    DailyBillingSummary:
      type: object
      properties:
        date:
          type: string
          format: date
        totalRevenue:
          type: number
        totalInvoices:
          type: integer
        paidInvoices:
          type: integer
        pendingInvoices:
          type: integer
        cashReceipts:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
      required: [date, totalRevenue, totalInvoices, paidInvoices, pendingInvoices, cashReceipts]
    BillingReconciliation:
      type: object
      properties:
        date:
          type: string
          format: date
        summary:
          type: object
          properties:
            collectedTotal:   { type: number }
            collectedCount:   { type: integer }
            invoicedTotal:    { type: number }
            invoicedCount:    { type: integer }
            outstandingTotal: { type: number }
            outstandingCount: { type: integer }
            cancelledTotal:   { type: number }
            cancelledCount:   { type: integer }
          required: [collectedTotal, collectedCount, invoicedTotal, invoicedCount, outstandingTotal, outstandingCount, cancelledTotal, cancelledCount]
        payments:
          type: array
          items:
            type: object
            properties:
              id:            { type: integer }
              invoiceNumber: { type: string }
              patientName:   { type: string, nullable: true }
              total:         { type: number }
              paidAt:        { type: string, format: date-time, nullable: true }
              createdByName: { type: string, nullable: true }
            required: [id, invoiceNumber, total]
      required: [date, summary, payments]
...